Organize validator functions better (no functional change)
