+++ src/twilio_call_center/utils.py
@@ -1,0 +1,3 @@
+
+def split_csv_list(value):
+    return map(str.strip, value.split(','))
